requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
av==15.0.0
backoff==2.2.1
beautifulsoup4==4.13.4
browser-use==0.5.11
bubus==1.5.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.6
ctranslate2==4.6.0
distro==1.9.0
fastapi==0.116.1
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.29.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
markdown-pdf==1.5
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.13.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.3
onnxruntime==1.22.1
openai==1.99.9
orjson==3.11.2
packaging==25.0
pandas==2.3.1
patchright==1.52.5
playwright==1.54.0
portalocker==2.10.1
posthog==6.5.0
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
PyMuPDF==1.25.3
pyparsing==3.2.3
pypdf==6.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.27.0
rsa==4.9.1
screeninfo==0.8.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
websockets==15.0.1
zstandard==0.24.0
